+++ entrega2DPOO/src/usuario/OperadorAtraccion.java
@@ -9,7 +9,7 @@ public class OperadorAtraccion extends Empleado{
 	private String lugarDeTrabajoDia = null;
 	
 	
-	public OperadorAtraccion(String nombre, String login, String password, float altura, float peso, String rol) {
+	public OperadorAtraccion(String nombre, String login, String password, int altura, int peso, String rol) {
 		super(nombre, login, password, altura, peso, rol, null);
 		this.capacitaciones = new ArrayList<>();
 	}
