+++ entrega2DPOO/src/usuario/ServicioGeneral.java
@@ -2,7 +2,7 @@ package usuario;
 
 public class ServicioGeneral extends Empleado{
 	
-	public ServicioGeneral(String nombre, String login, String password, float altura, float peso, String rol
+	public ServicioGeneral(String nombre, String login, String password, int altura, int peso, String rol
 			, String lugarDeTrabajo) {
 		super(nombre, login, password, altura, peso, rol, lugarDeTrabajo);
 	}
